refactored transaction service
