+++ txpool/transaction_service.go
@@ -1,0 +1,5 @@
+package txpool
+
+type TransactionService interface {
+	TransferTxToLeader()
+}
